WC Marketplace Featured Image Fix
This was the elegant solution I was hoping for.

WCMP has a function that resolves a Media Library url to an attachment ID.  This function assumes that the url is in the local media library, which is why it was breaking when being fed a Google Storage url.

The WCMP function was not filterable, but it makes a call to `wp_get_upload_dir` which is filterable.  I intercepted the call in `functions.php` and replaced the local `baseurl` with the Google Storage url.

Works like a champ.

I’m only overriding the `baseurl` in the event that WP Stateless is active, and the code is being called in the admin.  Should be fine to overwrite, as WP Stateless takes over any scenario that would require the Media LIbrary’s baseurl.